бд в атласе

+++ js/inc/hw/server03/src/routes/post-router.ts
@@ -1,26 +1,33 @@
 import { Router, Request, Response } from "express";
 import {
-  IPostInputModel,
-  IPostViewModel,
   RequestWithBody,
   RequestWithBodyAndParams,
   RequestWithParams,
   Status,
 } from "../types";
+import { CreatePostInputModel } from "../models/input/post/create-post-input-model";
+import {
+  EnchancedPostOutputModel,
+  PostOutputModel,
+} from "../models/output/post-output-model";
 import { BlogRepository } from "../repositories/blog-repository";
 import { PostRepository } from "../repositories/post-repository";
 import { authMiddleware } from "../middlewares/auth-middleware";
-import { IPostDb } from "../db/db";
+import { PostDbType } from "../models/db/post-db";
 import { postValidation } from "../validators/post-validators";
 import { inputValidationMiddleware } from "../middlewares/inputValidationMiddleware";
+import { Params } from "./blog-router";
+import { ObjectId } from "mongodb";
+import { UpdatePostInputModel } from "../models/input/post/update-post-input-model";
+import { CreateBlogModel } from "../models/input/blog/create-blog-input-model";
 
 export const postRouter = Router();
 
-postRouter.get("/", (req: Request, res: Response) => {
-  const dbPosts = PostRepository.getAllPosts();
-  const dbBlogs = BlogRepository.getAllBlogs();
-  const posts: IPostViewModel[] = dbPosts.map((post) => {
-    const blog = dbBlogs.find((blog) => blog.id === post.blogId);
+postRouter.get("/", async (req: Request, res: Response) => {
+  const dbPosts = await PostRepository.getAllPosts();
+  const dbBlogs = await BlogRepository.getAllBlogs();
+  const posts: EnchancedPostOutputModel[] = dbPosts.map((post) => {
+    const blog = dbBlogs.find((blog) => blog.id.toString() === post.blogId);
     return {
       ...post,
       blogName: blog ? blog.name : "",
@@ -31,15 +38,24 @@ postRouter.get("/", (req: Request, res: Response) => {
 
 postRouter.get(
   "/:id",
-  (req: RequestWithParams<{ id: string }>, res: Response<IPostViewModel>) => {
-    const post = PostRepository.getPostById(req.params.id);
+  async (
+    req: RequestWithParams<Params>,
+    res: Response<EnchancedPostOutputModel>
+  ) => {
+    const id = req.params.id;
 
+    if (!ObjectId.isValid(id)) {
+      res.sendStatus(Status.NotFound_404);
+      return;
+    }
+
+    const post = await PostRepository.getPostById(id);
     if (!post) {
       res.sendStatus(Status.NotFound_404);
       return;
     }
 
-    const blog = BlogRepository.getBlogById(post.blogId);
+    const blog = await BlogRepository.getBlogById(post.blogId);
     const viewPost = { ...post, blogName: blog ? blog.name : "" };
     res.status(Status.Ok_200).json(viewPost);
   }
@@ -50,22 +66,29 @@ postRouter.post(
   authMiddleware,
   postValidation(),
   inputValidationMiddleware,
-  (req: RequestWithBody<IPostInputModel>, res: Response<IPostViewModel>) => {
+  async (
+    req: RequestWithBody<CreatePostInputModel>,
+    res: Response<PostOutputModel>
+  ) => {
     const { title, content, shortDescription, blogId } = req.body;
 
-    const newPostDb: IPostDb = {
-      id: new Date().getTime().toString(),
+    const newPost: CreatePostInputModel = {
       title,
       content,
       shortDescription,
       blogId,
     };
 
-    PostRepository.createPost(newPostDb);
+    const postId = await PostRepository.createPost(newPost);
+    const post = await PostRepository.getPostById(postId);
+    if (!post) {
+      res.sendStatus(Status.NotFound_404);
+      return;
+    }
 
-    const blog = BlogRepository.getBlogById(blogId);
-    const newPostView: IPostViewModel = {
-      ...newPostDb,
+    const blog = await BlogRepository.getBlogById(blogId);
+    const newPostView: EnchancedPostOutputModel = {
+      ...post,
       blogName: blog ? blog.name : "",
     };
 
@@ -78,30 +101,36 @@ postRouter.put(
   authMiddleware,
   postValidation(),
   inputValidationMiddleware,
-  (
-    req: RequestWithBodyAndParams<{ id: string }, IPostInputModel>,
-    res: Response<IPostViewModel>
+  async (
+    req: RequestWithBodyAndParams<{ id: string }, CreatePostInputModel>,
+    res: Response<PostOutputModel>
   ) => {
+    const id = req.params.id;
+
+    if (!ObjectId.isValid(id)) {
+      res.sendStatus(Status.NotFound_404);
+      return;
+    }
+
     const { title, content, shortDescription, blogId } = req.body;
 
-    const updatedPost: IPostInputModel = {
+    const updatedPost: UpdatePostInputModel = {
       title,
       content,
       shortDescription,
       blogId,
     };
 
-    PostRepository.updatePost(req.params.id, updatedPost);
-
-    const blog = BlogRepository.getBlogById(blogId);
-    const post = PostRepository.getPostById(req.params.id);
+    await PostRepository.updatePost(id, updatedPost);
+    const blog = await BlogRepository.getBlogById(blogId);
+    const post = await PostRepository.getPostById(id);
 
     if (!post) {
       res.sendStatus(Status.NotFound_404);
       return;
     }
 
-    const newPostView: IPostViewModel = {
+    const newPostView: EnchancedPostOutputModel = {
       ...post,
       blogName: blog ? blog.name : "",
     };
@@ -113,13 +142,24 @@ postRouter.put(
 postRouter.delete(
   "/:id",
   authMiddleware,
-  (req: RequestWithParams<{ id: string }>, res) => {
-    const post = PostRepository.getPostById(req.params.id);
+  async (req: RequestWithParams<Params>, res) => {
+    const id = req.params.id;
+
+    if (!ObjectId.isValid(id)) {
+      res.sendStatus(Status.NotFound_404);
+      return;
+    }
+
+    const post = await PostRepository.getPostById(id);
     if (!post) {
       res.sendStatus(Status.NotFound_404);
     }
 
-    PostRepository.deletePost(req.params.id);
+    const isDeleted = await PostRepository.deletePost(req.params.id);
+    if (!isDeleted) {
+      res.sendStatus(Status.NotFound_404);
+      return;
+    }
     res.sendStatus(Status.NoContent_204);
   }
 );
